pyproject.toml
[project]
name = "oktoberfest-monitor"
version = "0.1.0"
description = "Monitors Oktoberfest tent reservation pages and pushes iPhone notifications via Pushover when 25.09.2026 / 26.09.2026 availability changes."
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "selectolax>=0.3.21",
    "pyyaml>=6.0",
    "pydantic>=2.7",
    "jsonpath-ng>=1.6",
]

[project.optional-dependencies]
headless = ["playwright>=1.45"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
